+++ flexmeasures/api/v3_0/assets.py
@@ -70,7 +70,6 @@ from flexmeasures.api.common.responses import (
     request_processed,
 )
 from flexmeasures.api.common.schemas.users import AccountIdField
-from flexmeasures.data.config import most_recent_beliefs_mview
 from flexmeasures.api.common.schemas.assets import default_response_fields
 from flexmeasures.ui.utils.view_utils import clear_session, set_session_variables
 from flexmeasures.auth.policy import check_access, user_has_admin_access
@@ -962,6 +961,7 @@ class AssetAPI(FlaskView):
             ),
             "most_recent_beliefs_only": fields.Boolean(required=False),
             "use_materialized_view": fields.Boolean(required=False, load_default=True),
+            "include_live_tail": fields.Boolean(required=False, load_default=None),
             "compress_json": fields.Boolean(required=False),
         },
         location="query",
@@ -1003,7 +1003,6 @@ class AssetAPI(FlaskView):
             - Assets
         """
         sensors = SensorsToShowSchema.flatten(asset.validate_sensors_to_show())
-        kwargs["most_recent_beliefs_mview"] = most_recent_beliefs_mview
         return asset.chart_data_json(sensors=sensors, **kwargs)
 
     @route("/<id>/auditlog")