+++ flexmeasures/data/models/generic_assets.py
@@ -7,7 +7,7 @@ import json
 from flask import current_app
 from flask_security import current_user
 import pandas as pd
-from sqlalchemy import select, and_, Table
+from sqlalchemy import select, and_
 from sqlalchemy.ext.hybrid import hybrid_method
 from sqlalchemy.sql.expression import func
 from sqlalchemy.ext.mutable import MutableDict, MutableList
@@ -1129,7 +1129,7 @@ class GenericAsset(db.Model, AuthModelMixin):
         compress_json: bool = False,
         resolution: timedelta | None = None,
         use_materialized_view: bool = True,
-        most_recent_beliefs_mview: Table | None = None,
+        include_live_tail: bool | None = None,
     ) -> BeliefsDataFrame | str:
         """Search all beliefs about events for all sensors of this asset
 
@@ -1175,7 +1175,7 @@ class GenericAsset(db.Model, AuthModelMixin):
                 one_deterministic_belief_per_event_per_source=True,
                 resolution=resolution,
                 use_materialized_view=use_materialized_view,
-                most_recent_beliefs_mview=most_recent_beliefs_mview,
+                include_live_tail=include_live_tail,
             )
         if as_json and not compress_json:
             from flexmeasures.data.services.time_series import simplify_index